dbus v0.6.0版初次发布: # 重要功能优化 * 优化dbus打包,添加deployer打包工程； * 优化dbus部署,所有配置统一在conf/config.properties中配置； * 全量部分代码重构,优化mysql拉取逻辑,防止出现oom导致worker自动重启导致任务拉取失败； * 新增sinker模块,实现增量数据落hdfs；
# 功能新增
* 新增全量断点续传功能,针对分片数目较多的任务,可以续传未完成的分片任务；
* 自动部署ogg、canal：
  新增自动部署canal,自动添加ogg目标端replicat进程功能；
  新增ogg、canal自动部署、重新部署、自动漂移功能；
  新增集群服务器canal、ogg部署详情集中管理功能；
* 新增批量重启拓扑功能,方便批量升级dbus mysql-extractor,dispatcher-appender、splitter-puller；
* 新增数据线一键检测功能，包括拓扑、ogg、canal、主备延时状态自动获取，此功能目前只提供给心跳程序使用；
* 重新设计优化，拓扑停止、状态获取,不再依赖stormUI；
* 新增更新表结构功能，用于加表后version为null，源端和dbus表结构不一致时自动更新表结构；

+++ dbus-keeper/keeper-service/src/main/java/com/creditease/dbus/domain/mapper/DataSchemaMapper.java
@@ -2,7 +2,7 @@
  * <<
  * DBus
  * ==
- * Copyright (C) 2016 - 2018 Bridata
+ * Copyright (C) 2016 - 2019 Bridata
  * ==
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -18,6 +18,7 @@
  * >>
  */
 
+
 package com.creditease.dbus.domain.mapper;
 
 import com.creditease.dbus.domain.model.DataSchema;
@@ -27,19 +28,19 @@ import java.util.List;
 import java.util.Map;
 
 public interface DataSchemaMapper {
-    List<Map<String,Object>> searchSchemaAndDs(Map<String, Object> param);
+    List<Map<String, Object>> searchSchemaAndDs(Map<String, Object> param);
 
-    List<DataSchema> searchSchema(@Param("schemaId") Long schemaId, @Param("dsId") Long dsId, @Param("schemaName") String schemaName);
+    List<DataSchema> searchSchema(@Param("schemaId") Integer schemaId, @Param("dsId") Integer dsId, @Param("schemaName") String schemaName);
 
     int insert(DataSchema record);
 
     int update(DataSchema record);
 
-    int deleteBySchemaId(Long id);
+    int deleteBySchemaId(Integer id);
 
     int updateSchemaStatusByPrimaryKey(Map<String, Object> param);
 
-    DataSchema selectById(Long id);
+    DataSchema selectById(Integer id);
 
     DataSchema findByDsIdAndSchemaName(@Param("dsId") Integer dsId, @Param("schemaName") String schemaName);
